Move util exports to root
