+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2025-02-09
+
+### Changed
+
+- Moved export of `aliasColors` util
+
 ## [0.1.6] - 2025-02-09
 
 ### Fixed
