+++ src/public/index.ts
@@ -1,0 +1,2 @@
+export * from './alias';
+export * from './types-public';
